requirements.txt
aiosqlite==0.22.1
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.1
attrs==25.4.0
authlib==1.6.6
backoff==2.2.1
bracex==2.6
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cryptography==46.0.3
deepagents==0.3.1
distro==1.9.0
docstring-parser==0.17.0
fastapi==0.123.10
filetype==1.2.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.132.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.39.0
google-cloud-bigquery-storage==2.36.0
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.13.0
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.61.0
google-cloud-speech==2.35.0
google-cloud-storage==3.7.0
google-cloud-trace==1.17.0
google-crc32c==1.8.0
google-genai==1.56.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib-metadata==8.7.1
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.0
langchain-anthropic==1.3.1
langchain-core==1.2.6
langchain-google-genai==4.1.3
langchain-openai==1.1.6
langchain-sandbox==0.0.6
langfuse==3.11.2
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.1
mako==1.3.10
markupsafe==3.0.3
mcp==1.25.0
mmh3==5.2.0
numpy==2.4.0
openai==2.14.0
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
proto-plus==1.27.0
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pyjwt==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pyyaml==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.45
sqlalchemy-spanner==1.17.2
sqlparse==0.5.5
sse-starlette==3.1.2
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid-utils==0.12.0
uvicorn==0.40.0
watchdog==6.0.0
wcmatch==10.1
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
